Rename commands to mutations
