+++ lib/templates/mutations/deleteTemplate.ts
