+++ lib/templates/mutations/removeAssignedUserFromTemplate.ts
