+++ lib/templates/mutations/syncAssignedUsersForTemplate.ts
