+++ lib/templates/mutations/updateClosedData.ts
