+++ lib/templates/mutations/updateIsPublishedForTemplate.ts
