+++ lib/templates/mutations/updateSecurityAttribute.ts
